# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/historytaking": {
            "get": {
                "description": "list historytaking entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List historytaking entities",
                "operationId": "list-historytaking",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Historytaking"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create historytaking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create historytaking",
                "operationId": "create-historytaking",
                "parameters": [
                    {
                        "description": "Historytaking entity",
                        "name": "historytaking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Historytaking"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Historytaking"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/historytaking/{id}": {
            "get": {
                "description": "get historytaking by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a historytaking entity by ID",
                "operationId": "get-historytaking",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Historytaking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Historytaking"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/medicalrecordstaff": {
            "get": {
                "description": "list medicalrecordstaff entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List medicalrecordstaff entities",
                "operationId": "list-medicalrecordstaff",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Medicalrecordstaff"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/medicalrecordstaff/{id}": {
            "get": {
                "description": "get medicalrecordstaff by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a medicalrecordstaff entity by ID",
                "operationId": "get-medicalrecordstaff",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Medicalrecordstaff ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Medicalrecordstaff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Historytaking": {
            "type": "object",
            "properties": {
                "bp": {
                    "type": "string"
                },
                "datetime": {
                    "type": "string"
                },
                "department": {
                    "type": "integer"
                },
                "hight": {
                    "type": "string"
                },
                "nurse": {
                    "type": "integer"
                },
                "oxygen": {
                    "type": "string"
                },
                "patientrecord": {
                    "type": "integer"
                },
                "pulse": {
                    "type": "string"
                },
                "respiration": {
                    "type": "string"
                },
                "symptom": {
                    "type": "string"
                },
                "symptomseverity": {
                    "type": "integer"
                },
                "temp": {
                    "type": "string"
                },
                "weight": {
                    "type": "string"
                }
            }
        },
...
        "ent.Department": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department holds the value of the \"department\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DepartmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DepartmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DepartmentEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDepartment2doctorinfo": {
                    "description": "EdgesOfDepartment2doctorinfo holds the value of the EdgesOfDepartment2doctorinfo edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Doctorinfo"
                    }
                },
                "EdgesOfHistorytaking": {
                    "description": "EdgesOfHistorytaking holds the value of the EdgesOfHistorytaking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Historytaking"
                    }
                }
            }
        },
...
        "ent.Historytaking": {
            "type": "object",
            "properties": {
                "bp": {
                    "description": "Bp holds the value of the \"bp\" field.",
                    "type": "integer"
                },
                "datetime": {
                    "description": "Datetime holds the value of the \"datetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the HistorytakingQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.HistorytakingEdges"
                },
                "hight": {
                    "description": "Hight holds the value of the \"hight\" field.",
                    "type": "number"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "oxygen": {
                    "description": "Oxygen holds the value of the \"oxygen\" field.",
                    "type": "string"
                },
                "pulse": {
                    "description": "Pulse holds the value of the \"pulse\" field.",
                    "type": "integer"
                },
                "respiration": {
                    "description": "Respiration holds the value of the \"respiration\" field.",
                    "type": "integer"
                },
                "symptom": {
                    "description": "Symptom holds the value of the \"symptom\" field.",
                    "type": "string"
                },
                "temp": {
                    "description": "Temp holds the value of the \"temp\" field.",
                    "type": "number"
                },
                "weight": {
                    "description": "Weight holds the value of the \"weight\" field.",
                    "type": "number"
                }
            }
        },
        "ent.HistorytakingEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDepartment": {
                    "description": "EdgesOfDepartment holds the value of the EdgesOfDepartment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "EdgesOfNurse": {
                    "description": "EdgesOfNurse holds the value of the EdgesOfNurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "EdgesOfPatientrecord": {
                    "description": "EdgesOfPatientrecord holds the value of the EdgesOfPatientrecord edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patientrecord"
                },
                "EdgesOfSymptomseverity": {
                    "description": "EdgesOfSymptomseverity holds the value of the EdgesOfSymptomseverity edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Symptomseverity"
                }
            }
        },
...
        "ent.Medicalrecordstaff": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicalrecordstaffQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicalrecordstaffEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.MedicalrecordstaffEdges": {
            "type": "object",
            "properties": {
                "EdgesOfMedicalrecordstaffPatientrights": {
                    "description": "EdgesOfMedicalrecordstaffPatientrights holds the value of the EdgesOfMedicalrecordstaffPatientrights edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patientrights"
                    }
                },
                "EdgesOfPatientrecord": {
                    "description": "EdgesOfPatientrecord holds the value of the EdgesOfPatientrecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patientrecord"
                    }
                },
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Nurse": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NurseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.NurseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "nursinglicense": {
                    "description": "Nursinglicense holds the value of the \"nursinglicense\" field.",
                    "type": "string"
                },
                "position": {
                    "description": "Position holds the value of the \"position\" field.",
                    "type": "string"
                }
            }
        },
        "ent.NurseEdges": {
            "type": "object",
            "properties": {
                "EdgesOfHistorytaking": {
                    "description": "EdgesOfHistorytaking holds the value of the EdgesOfHistorytaking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Historytaking"
                    }
                },
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
...
        "ent.Patientrecord": {
            "type": "object",
            "properties": {
                "Age": {
                    "description": "Age holds the value of the \"Age\" field.",
                    "type": "integer"
                },
                "Allergic": {
                    "description": "Allergic holds the value of the \"Allergic\" field.",
                    "type": "string"
                },
                "Bloodtype": {
                    "description": "Bloodtype holds the value of the \"Bloodtype\" field.",
                    "type": "string"
                },
                "Date": {
                    "description": "Date holds the value of the \"Date\" field.",
                    "type": "string"
                },
                "Disease": {
                    "description": "Disease holds the value of the \"Disease\" field.",
                    "type": "string"
                },
                "Email": {
                    "description": "Email holds the value of the \"Email\" field.",
                    "type": "string"
                },
                "Home": {
                    "description": "Home holds the value of the \"Home\" field.",
                    "type": "string"
                },
                "Idcardnumber": {
                    "description": "Idcardnumber holds the value of the \"Idcardnumber\" field.",
                    "type": "integer"
                },
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "Phonenumber": {
                    "description": "Phonenumber holds the value of the \"Phonenumber\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientrecordQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientrecordEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Patientrights": {
            "type": "object",
            "properties": {
                "PermissionDate": {
                    "description": "PermissionDate holds the value of the \"PermissionDate\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientrightsQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientrightsEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "insurance_id": {
                    "type": "integer"
                },
                "patientrightstype_id": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Symptomseverity": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SymptomseverityQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SymptomseverityEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "symptomseverity": {
                    "description": "Symptomseverity holds the value of the \"symptomseverity\" field.",
                    "type": "string"
                }
            }
        },
        "ent.SymptomseverityEdges": {
            "type": "object",
            "properties": {
                "EdgesOfHistorytaking": {
                    "description": "EdgesOfHistorytaking holds the value of the EdgesOfHistorytaking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Historytaking"
                    }
                }
            }
        },
...
        "ent.User": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "images": {
                    "description": "Images holds the value of the \"images\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}